#2 게시판 작업(JPA) post -> posts 병합 완료
