--- src/main/java/org/iclass/board/dao/PostMapper.java
@@ -1,9 +1,0 @@
-package org.iclass.board.dao;
-
-import org.apache.ibatis.annotations.Mapper;
-import org.iclass.board.dto.PostDTO;
-
-@Mapper
-public interface PostMapper {
-    void savePost(PostDTO post);
-}
