+++ src/main/java/org/iclass/board/dao/PostsMapper.java
@@ -1,0 +1,9 @@
+package org.iclass.board.dao;
+
+import org.apache.ibatis.annotations.Mapper;
+import org.iclass.board.dto.PostsDTO;
+
+@Mapper
+public interface PostsMapper {
+    void savePost(PostsDTO post);
+}
