+++ src/main/java/org/iclass/board/dto/UserDTO.java
@@ -14,7 +14,7 @@ public class UserDTO {
     private String username;
     private String password;
     private String email;
-//    private String userRole;
+    private String userRole;
     private Date createdAt;
     private String isActive;
 }
