+++ src/main/java/org/iclass/board/service/PostsService.java
@@ -3,6 +3,7 @@ package org.iclass.board.service;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.iclass.board.dao.PostsMapper;
 import org.iclass.board.dto.PostsDTO;
 import org.iclass.board.entity.PostsEntity;
 import org.iclass.board.repository.PostsRepository;
@@ -17,6 +18,15 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class PostsService {
 
+    /* myBatis 구간 */
+    private final PostsMapper postsMapper;
+
+    public void savePost(PostsDTO post) {
+        postsMapper.savePost(post);
+    }
+
+
+    /* JPA 구간 */
     private final PostsRepository postsRepository;
 
     public List<PostsDTO> list() {
